[FIX] Empty profile list -> import default profile
Prior to this commit, if the loaded profile list was empty (or did not
exist), the default profile path was automatically added to the profiles
list.  However, plugin dependencies were not installed.

As of this commit, instead of simply adding the default profile path to
the profiles list, the default profile path is **imported**. This way
plugin dependencies are installed and the version of the profile is
verified against the installed **major version** of MicroDrop.