Upgrade `microdrop-launcher` package in background
Prior to this commit, despite using the `ThreadPoolExecutor` from the
`concurrent.futures` package, the `auto_upgrade` function was not being
run concurrently with the MicroDrop Launcher dialog.  It seems that the
`auto_upgrade` thread was blocking unexpectedly.

As of this commit, the automatic upgrade is executed via a
`subprocess.Popen` call, which actually runs concurrently with the
MicroDrop Launcher dialog.