Fix MicroDrop upgrades: use Conda versions
Prior to this commit, `pkg_resources` was used to query the installed
version of MicroDrop.  As of this commit, the installed MicroDrop version
is determined using `conda_helpers.package_version` function.

Related to this change, this commit adds the
`profile.check_version_cache_for_upgrade` function to compare the
installed MicroDrop version to the latest cached available version and
offer the user to upgrade if a new version is available.

Note that this commit requires `conda-helpers >=0.9` and the Conda build
recipe has been updated accordingly.